requirements.txt
paho-mqtt
schedule
gpiozero
RPi.GPIO
picamera2
adafruit-circuitpython-dht
pydrive2
